+++ jest.config.js
@@ -1,0 +1,15 @@
+module.exports = {
+  preset: "ts-jest",
+  transform: {
+    "^.+\\.ts?$": [
+      "ts-jest",
+      {
+        tsconfig: "tsconfig.tests.json",
+      },
+    ],
+  },
+  testEnvironment: "node",
+  roots: ["<rootDir>/tests"],
+  testMatch: ["**/test*.ts"],
+  collectCoverageFrom: ["src/**/*.{ts,js}", "!src/**/*.d.ts"],
+};