+++ src/handlers/hover.ts
@@ -1,66 +1,99 @@
 import {
   Connection,
   Hover,
-  Position,
   TextDocuments,
   TextDocumentPositionParams,
   Range,
 } from "vscode-languageserver";
 import { TextDocument } from "vscode-languageserver-textdocument";
 
-import { TodotxtTokenType } from "../parser/tokenTypes";
-import {
-  IndexedToken,
-  tokenizeLine,
-  getTokenAtPosition,
-} from "../parser/tokenizer";
-import {
-  determineTodotxtTokenType,
-  getDescriptionStart,
-} from "../parser/lexer";
-import { getRows } from "../parser/utils";
+import { Token, TodotxtTokenType } from "../parser/tokenTypes";
+import { decodeTokenType, encodeTokenType } from "../parser/utils";
+import { retrieveDocTokens, getIndexAtPosition, getTokenEnd } from "../tokenManager";
 
 const TODOTXT_DOC_URL =
   "https://github.com/todotxt/todo.txt?tab=readme-ov-file";
-const tokenTypeDocUrlMap: Record<TodotxtTokenType, string> = {
-  completionMark: `${TODOTXT_DOC_URL}#complete-tasks-2-format-rules`,
-  priority: `${TODOTXT_DOC_URL}#rule-1-if-priority-exists-it-always-appears-first`,
-  creationDate: `${TODOTXT_DOC_URL}#rule-2-a-tasks-creation-date-may-optionally-appear-directly-after-priority-and-a-space`,
-  completionDate: `${TODOTXT_DOC_URL}#rule-2-the-date-of-completion-appears-directly-after-the-x-separated-by-a-space`,
-  project: `${TODOTXT_DOC_URL}#project`,
-  context: `${TODOTXT_DOC_URL}#context`,
-  keyValue: `${TODOTXT_DOC_URL}#additional-file-format-definitions`,
-  description: `${TODOTXT_DOC_URL}#todotxt-format-rules`,
-};
+const TOKEN_TYPE_DOC_URL_MAP: string[] = [
+  `${TODOTXT_DOC_URL}#todotxt-format-rules`,
+  `${TODOTXT_DOC_URL}#rule-1-if-priority-exists-it-always-appears-first`,
+  `${TODOTXT_DOC_URL}#rule-2-a-tasks-creation-date-may-optionally-appear-directly-after-priority-and-a-space`,
+  `${TODOTXT_DOC_URL}#rule-2-the-date-of-completion-appears-directly-after-the-x-separated-by-a-space`,
+  `${TODOTXT_DOC_URL}#complete-tasks-2-format-rules`,
+  `${TODOTXT_DOC_URL}#project`,
+  `${TODOTXT_DOC_URL}#context`,
+  `${TODOTXT_DOC_URL}#additional-file-format-definitions`,
+];
 
+// TODO: more consistent function prototype
 const createHoverContent = (
-  token: string,
-  tokenType: TodotxtTokenType,
+   tokenType: number, tokenTypeName: TodotxtTokenType, content: string
 ): string => {
+  // TODO: priorities, creation and completion dates and completion marks are not highlighted (by tree-sitter) because "(tokenType) " is prepending. Think about a solution.
   return `\`\`\`todo.txt
-(${tokenType}) ${token}
+(${tokenTypeName}) ${content}
 \`\`\`
 ___
-[Format spec](${tokenTypeDocUrlMap[tokenType]})`;
+[Format spec](${TOKEN_TYPE_DOC_URL_MAP[tokenType]})`;
 };
 
-const determineTokenRange = (
-  token: IndexedToken,
-  tokenType: TodotxtTokenType,
-  rows: string[],
-  position: Position,
-): Range => {
-  let start: number, end: number;
-  if (tokenType !== "description") {
-    start = token.start;
-    end = token.start + token.token.length;
-  } else {
-    start = getDescriptionStart(rows[position.line]);
-    end = rows[position.line].length;
+const BEGINNING_TOKEN_PATTERNS: number[] = [
+  "priority",
+  "completionMark",
+  "creationDate",
+  "completionDate"
+].map((tokenType: string) => encodeTokenType(tokenType));
+
+const getCoverageInterval = (tokens: Token[], index: number): [number, number] => {
+  let start: number = index; let end: number = index;
+
+  if (decodeTokenType(tokens[index].tokenType) === "description") {
+    for (;
+      start > 0
+      && tokens[start - 1]?.line === tokens[index].line
+      && BEGINNING_TOKEN_PATTERNS.indexOf(tokens[start - 1].tokenType) === -1;
+      start--
+    ) {}
+
+    for (;
+      end < tokens.length - 1
+      && tokens[end + 1]?.line === tokens[index].line;
+      end++
+    ) {}
   }
+
+  return [start, end];
+}
+
+const getTokenIntervalText = (
+	tokens: Token[],
+ 	intervalStart: number,
+  intervalEnd: number
+): string => {
+  let resultText: string = tokens[intervalStart].content;
+  for (let i: number = intervalStart + 1; i <= intervalEnd; i++) {
+ 	/* NOTE: tokenizing text doesn't count which exact whitespaces split tokens.
+  	Task description inside hover response may be incorrect if \t are used inside task text. */
+    resultText += " ".repeat(
+    	tokens[i].character - getTokenEnd(tokens[i-1])
+    ) + tokens[i].content;
+  }
+	return resultText;
+};
+
+const getTokenRange = (
+  tokens: Token[],
+  tokenIntervalStart: number,
+  tokenIntervalEnd: number,
+): Range => {
   return {
-    start: { line: position.line, character: start } as Position,
-    end: { line: position.line, character: end } as Position,
+    start: {
+    	line: tokens[tokenIntervalStart].line,
+     	character: tokens[tokenIntervalStart].character
+    },
+    end: {
+    	line: tokens[tokenIntervalEnd].line,
+    	character: tokens[tokenIntervalEnd].character + tokens[tokenIntervalEnd].content.length
+    },
   };
 };
 
@@ -72,39 +105,38 @@ export const registerHoverHandler = (
     const doc = documents.get(params.textDocument.uri);
     if (!doc) return null;
 
-    const rows: string[] = getRows(doc.getText());
-    const row: string = rows[params.position.line];
+		const docTokens = retrieveDocTokens(doc);
+		if (!docTokens) return null;
 
-    const tokens: IndexedToken[] = tokenizeLine(row);
-    const token: IndexedToken = getTokenAtPosition(
-      tokens,
-      params.position.character,
-    );
-    if (!token) return null;
-
-    let tokenType: TodotxtTokenType = determineTodotxtTokenType(token, tokens);
-    const tokenRange: Range = determineTokenRange(
-      token,
-      tokenType,
-      rows,
+    const tokenIndex: number = getIndexAtPosition(
+      docTokens,
       params.position,
-    );
+      false
+		);
+    // connection.console.debug(`index at position: ${tokenIndex}`);
+    const currentToken: Token = docTokens[tokenIndex];
+    if (!currentToken) return null;
 
-    const content: string = createHoverContent(
-      tokenType === "description"
-        ? rows[params.position.line].slice(
-            getDescriptionStart(rows[params.position.line]),
-          )
-        : token.token,
-      tokenType,
-    );
+    const tokenTypeName: TodotxtTokenType = decodeTokenType(currentToken.tokenType);
+		const [tokenIntervalStart, tokenIntervalEnd]: [number, number] = getCoverageInterval(
+			docTokens, tokenIndex
+		);
 
     return {
       contents: {
         kind: "markdown",
-        value: content,
+        value: createHoverContent(
+          currentToken.tokenType,
+           tokenTypeName,
+           `${tokenIntervalStart} ${tokenIntervalEnd}`
+          	// tokenTypeName === "description"
+           	// ? getTokenIntervalText(docTokens, tokenIntervalStart, tokenIntervalEnd)
+            // : currentToken.content
+        )
       },
-      range: tokenRange,
+      range: getTokenRange(
+        docTokens, tokenIntervalStart, tokenIntervalEnd
+      )
     };
   });
 };