+++ src/handlers/initialize.ts
@@ -1,45 +1,46 @@
 import {
-  InitializeParams,
-  InitializeResult,
-  TextDocumentSyncKind,
-  Connection,
+	InitializeParams,
+	InitializeResult,
+	TextDocumentSyncKind,
+	Connection,
 } from "vscode-languageserver/node";
 
 import { TodotxtTokenTypes } from "../parser/tokenTypes";
-import { NAME, VERSION } from "../version";
+import { NAME, VERSION } from "../info";
 
 export const registerInitializeHandler = (connection: Connection) => {
-  connection.onInitialize((params: InitializeParams): InitializeResult => {
-    const capabilities = params.capabilities;
+	connection.onInitialize((params: InitializeParams): InitializeResult => {
+		const capabilities = params.capabilities;
 
-    const initResult: InitializeResult = {
-      capabilities: {
-        textDocumentSync: TextDocumentSyncKind.Incremental,
-        hoverProvider: true,
-        semanticTokensProvider: {
-          legend: {
-            tokenTypes: [...TodotxtTokenTypes],
-            tokenModifiers: [],
-          },
-          full: true,
-        },
-        completionProvider: {
-          triggerCharacters: ["@", "+"],
-        },
-      },
-      serverInfo: {
-        name: NAME,
-        version: VERSION,
-      },
-    };
+		const initResult: InitializeResult = {
+			capabilities: {
+				textDocumentSync: TextDocumentSyncKind.Incremental,
+				hoverProvider: true,
+				semanticTokensProvider: {
+					legend: {
+						tokenTypes: [...TodotxtTokenTypes],
+						tokenModifiers: [],
+					},
+					full: true,
+				},
+				completionProvider: {
+					triggerCharacters: ["@", "+", ":", "x"],
+					// "allCommitCharacters": [ " " ],
+				},
+			},
+			serverInfo: {
+				name: NAME,
+				version: VERSION,
+			},
+		};
 
-    if (!!(capabilities.workspace && capabilities.workspace.workspaceFolders))
-      initResult.capabilities.workspace = {
-        workspaceFolders: {
-          supported: true,
-        },
-      };
+		if (!!(capabilities.workspace && capabilities.workspace.workspaceFolders))
+			initResult.capabilities.workspace = {
+				workspaceFolders: {
+					supported: true,
+				},
+			};
 
-    return initResult;
-  });
+		return initResult;
+	});
 };