+++ src/parser/tokenizer.ts
@@ -1,31 +1,44 @@
-export interface IndexedToken {
-  token: string;
-  start: number;
-}
+import { determineTodotxtTokenType } from "./lexer";
+import { encodeTokenType } from "./utils";
+import { Token } from "./tokenTypes";
 
 const tokenPattern: RegExp = /\S+/g;
 
-export const tokenizeLine = (text: string): IndexedToken[] => {
-  const indexexTokens: IndexedToken[] = [];
-  let match: RegExpExecArray | null;
+export const tokenizeLine = (
+	text: string,
+	targetLine: number = 0,
+	charOffset: number = 0,
+): Token[] => {
+	const tokens: Token[] = [];
+	let match: RegExpExecArray | null;
 
-  while ((match = tokenPattern.exec(text)) !== null) {
-    indexexTokens.push({
-      token: match[0],
-      start: match.index,
-    });
-  }
+	while ((match = tokenPattern.exec(text)) !== null) {
+		const newTokenChar: number = match.index + charOffset;
+		const newToken: Token = {
+			line: targetLine,
+			character: newTokenChar,
+			content: match[0],
+			tokenType: encodeTokenType(
+				determineTodotxtTokenType(match[0], targetLine, newTokenChar, tokens),
+			),
+			tokenModifiers: 0,
+		};
+		tokens.push(newToken);
+	}
 
-  return indexexTokens;
+	return tokens;
 };
 
-export const getTokenAtPosition = (
-  tokens: IndexedToken[],
-  offset: number,
-): IndexedToken => {
-  for (const token of tokens) {
-    if (offset >= token.start && offset <= token.start + token.token.length)
-      return token;
-  }
-  return null;
+const LINES_RE: RegExp = /\r?\n/;
+
+export const getTokenizedText = (
+	text: string,
+	lineOffset: number = 0,
+	charOffset: number = 0,
+): Token[] => {
+	const textLines: string[] = text.split(LINES_RE);
+	const tokens: Token[] = tokenizeLine(textLines[0], lineOffset, charOffset);
+	for (let i: number = 1; i < textLines.length; i++)
+		tokens.push(...tokenizeLine(textLines[i], lineOffset + i, 0));
+	return tokens;
 };