+++ src/parser/utils.ts
@@ -1,5 +1,11 @@
-const ROWS_RE: RegExp = /\r?\n/;
+import { legend, TodotxtTokenType, TodotxtTokenTypes } from "./tokenTypes";
 
-export const getRows = (text: string): string[] => {
-  return text.split(ROWS_RE);
+export const encodeTokenType = (type: TodotxtTokenType): number => {
+	let ind: number = TodotxtTokenTypes.indexOf(type);
+	if (ind === -1) throw new Error("Incorrect token type!");
+	return ind;
+};
+
+export const decodeTokenType = (tokenCode: number): TodotxtTokenType => {
+	return legend.tokenTypes[tokenCode];
 };