+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+0.3 (06.12.25)
+
+- Changed document token format, caching tokens, cache delta updates (textDocument/didChange) support.
+- Fixed various lexer errors.
+- Provided unit tests for text tokenizing and token position walking.
+- Improved completions for key:values. 
+
 0.2 (16.11.25)
 
 - Suggestions for +projects and @contexts.